fix: move options class
