+++ packages/sdk-ts/src/client/abacus/grpc/AbacusGrpcApi.ts
@@ -3,7 +3,7 @@ import * as PointsSvcPb from '@injectivelabs/abacus-proto-ts-v2/generated/points
 import { PointsSvcClient } from '@injectivelabs/abacus-proto-ts-v2/generated/points_svc_pb.client'
 import BaseGrpcConsumer from '../../base/BaseGrpcConsumer.js'
 import { AbacusGrpcTransformer } from './transformers/index.js'
-import type { GrpcWebTransportAdditionalOptions } from '../../../utils/grpc.js'
+import type { GrpcWebTransportAdditionalOptions } from '../../../types'
 
 export class AbacusGrpcApi extends BaseGrpcConsumer {
   protected module: string = IndexerErrorModule.Abacus
