+++ packages/sdk-ts/src/client/chain/grpc/ChainGrpcEvmApi.ts
@@ -3,7 +3,7 @@ import { QueryClient as InjectiveEvmV1QueryClient } from '@injectivelabs/core-pr
 import { ChainModule } from '../types/index.js'
 import BaseGrpcConsumer from '../../base/BaseGrpcConsumer.js'
 import { ChainGrpcEvmTransformer } from '../transformers/index.js'
-import type { GrpcWebTransportAdditionalOptions } from '../../../utils/grpc.js'
+import type { GrpcWebTransportAdditionalOptions } from '../../../types'
 
 /**
  * @category Chain Grpc API
