+++ packages/sdk-ts/src/client/chain/grpc/ChainGrpcGovApi.ts
@@ -6,7 +6,7 @@ import { ChainGrpcGovTransformer } from '../transformers/ChainGrpcGovTransformer
 import { ChainGrpcCommonTransformer } from '../transformers/ChainGrpcCommonTransformer.js'
 import type * as CosmosGovV1GovPb from '@injectivelabs/core-proto-ts-v2/generated/cosmos/gov/v1/gov_pb'
 import type { PaginationOption } from '../../../types/pagination.js'
-import type { GrpcWebTransportAdditionalOptions } from '../../../utils/grpc.js'
+import type { GrpcWebTransportAdditionalOptions } from '../../../types'
 
 /**
  * @category Chain Grpc API
