+++ packages/sdk-ts/src/client/chain/grpc/ChainGrpcInsuranceFundApi.ts
@@ -3,7 +3,7 @@ import { QueryClient as InjectiveInsuranceV1Beta1QueryClient } from '@injectivel
 import { ChainModule } from '../types/index.js'
 import BaseGrpcConsumer from '../../base/BaseGrpcConsumer.js'
 import { ChainGrpcInsuranceFundTransformer } from '../transformers/ChainGrpcInsuranceFundTransformer.js'
-import type { GrpcWebTransportAdditionalOptions } from '../../../utils/grpc.js'
+import type { GrpcWebTransportAdditionalOptions } from '../../../types'
 
 /**
  * @category Chain Grpc API
