+++ packages/sdk-ts/src/client/chain/grpc/ChainGrpcTendermintApi.ts
@@ -2,7 +2,7 @@ import * as CosmosBaseTendermintV1Beta1QueryPb from '@injectivelabs/core-proto-t
 import { ServiceClient as CosmosBaseTendermintV1Beta1QueryClient } from '@injectivelabs/core-proto-ts-v2/generated/cosmos/base/tendermint/v1beta1/query_pb.client'
 import { ChainModule } from '../types/index.js'
 import BaseGrpcConsumer from '../../base/BaseGrpcConsumer.js'
-import type { GrpcWebTransportAdditionalOptions } from '../../../utils/grpc.js'
+import type { GrpcWebTransportAdditionalOptions } from '../../../types'
 
 /**
  * @category Chain Grpc API
