+++ packages/sdk-ts/src/client/chain/grpc/ChainGrpcTxFeesApi.ts
@@ -3,7 +3,7 @@ import { QueryClient as InjectiveTxFeesV1Beta1QueryClient } from '@injectivelabs
 import { ChainModule } from '../types/index.js'
 import BaseGrpcConsumer from '../../base/BaseGrpcConsumer.js'
 import { ChainGrpcTxFeesTransformer } from '../transformers/index.js'
-import type { GrpcWebTransportAdditionalOptions } from '../../../utils/grpc.js'
+import type { GrpcWebTransportAdditionalOptions } from '../../../types'
 
 /**
  * @category Chain Grpc API
