+++ packages/sdk-ts/src/client/chain/grpc/ChainGrpcWasmXApi.ts
@@ -2,7 +2,7 @@ import * as InjectiveWasmxV1QueryPb from '@injectivelabs/core-proto-ts-v2/genera
 import { QueryClient as InjectiveWasmxV1QueryClient } from '@injectivelabs/core-proto-ts-v2/generated/injective/wasmx/v1/query_pb.client'
 import { ChainModule } from '../types/index.js'
 import BaseGrpcConsumer from '../../base/BaseGrpcConsumer.js'
-import type { GrpcWebTransportAdditionalOptions } from '../../../utils/grpc.js'
+import type { GrpcWebTransportAdditionalOptions } from '../../../types'
 
 /**
  * @category Chain Grpc API
