+++ packages/sdk-ts/src/client/indexer/grpc/IndexerGrpcAccountApi.ts
@@ -5,7 +5,7 @@ import { IndexerModule } from '../types/index.js'
 import { IndexerGrpcAccountTransformer } from '../transformers/index.js'
 import BaseIndexerGrpcConsumer from '../../base/BaseIndexerGrpcConsumer.js'
 import type { PaginationOption } from '../../../types/pagination.js'
-import type { GrpcWebTransportAdditionalOptions } from '../../../utils/grpc.js'
+import type { GrpcWebTransportAdditionalOptions } from '../../../types'
 
 /**
  * @category Indexer Grpc API
