+++ packages/sdk-ts/src/client/indexer/grpc/IndexerGrpcMegaVaultApi.ts
@@ -3,7 +3,7 @@ import { InjectiveMegavaultRPCClient } from '@injectivelabs/indexer-proto-ts-v2/
 import { IndexerModule } from '../types/index.js'
 import { IndexerGrpcMegaVaultTransformer } from '../transformers/index.js'
 import BaseIndexerGrpcConsumer from '../../base/BaseIndexerGrpcConsumer.js'
-import type { GrpcWebTransportAdditionalOptions } from '../../../utils/grpc.js'
+import type { GrpcWebTransportAdditionalOptions } from '../../../types'
 
 /**
  * @category Indexer Grpc API
