+++ packages/sdk-ts/src/client/indexer/grpc/IndexerGrpcMetaApi.ts
@@ -2,7 +2,7 @@ import * as InjectiveMetaRpcPb from '@injectivelabs/indexer-proto-ts-v2/generate
 import { InjectiveMetaRPCClient } from '@injectivelabs/indexer-proto-ts-v2/generated/injective_meta_rpc_pb.client'
 import { IndexerModule } from '../types/index.js'
 import BaseIndexerGrpcConsumer from '../../base/BaseIndexerGrpcConsumer.js'
-import type { GrpcWebTransportAdditionalOptions } from '../../../utils/grpc.js'
+import type { GrpcWebTransportAdditionalOptions } from '../../../types'
 
 /**
  * @category Indexer Grpc API
