+++ packages/sdk-ts/src/client/indexer/grpc/IndexerGrpcTradingApi.ts
@@ -3,7 +3,7 @@ import { InjectiveTradingRPCClient } from '@injectivelabs/indexer-proto-ts-v2/ge
 import { IndexerModule } from '../types/index.js'
 import BaseIndexerGrpcConsumer from '../../base/BaseIndexerGrpcConsumer.js'
 import type { MarketType, GridStrategyType } from '../types/index.js'
-import type { GrpcWebTransportAdditionalOptions } from '../../../utils/grpc.js'
+import type { GrpcWebTransportAdditionalOptions } from '../../../types'
 
 /**
  * @category Indexer Grpc API
