+++ packages/sdk-ts/src/client/indexer/grpc/IndexerGrpcWeb3GwApi.ts
@@ -10,7 +10,7 @@ import { stringToUint8Array } from '../../../utils/encoding.js'
 import { safeBigIntStringify } from '../../../utils/helpers.js'
 import { IndexerGrpcTransactionApi } from './IndexerGrpcTransactionApi.js'
 import type { EvmChainId, AccountAddress } from '@injectivelabs/ts-types'
-import type { GrpcWebTransportAdditionalOptions } from '../../../utils/grpc.js'
+import type { GrpcWebTransportAdditionalOptions } from '../../../types'
 
 /**
  * @category Indexer Grpc API
