+++ packages/sdk-ts/src/client/olp/grpc/OLPGrpcApi.ts
@@ -3,7 +3,7 @@ import * as DmmPb from '@injectivelabs/olp-proto-ts-v2/generated/dmm_pb'
 import { InjectiveDmmV2RPCClient } from '@injectivelabs/olp-proto-ts-v2/generated/dmm_pb.client'
 import { DmmGrpcTransformer } from './transformers/index.js'
 import BaseIndexerGrpcConsumer from '../../base/BaseIndexerGrpcConsumer.js'
-import type { GrpcWebTransportAdditionalOptions } from '../../../utils/grpc.js'
+import type { GrpcWebTransportAdditionalOptions } from '../../../types'
 
 export class OLPGrpcApi extends BaseIndexerGrpcConsumer {
   protected module: string = IndexerErrorModule.OLP
