+++ packages/sdk-ts/src/types/grpc.ts
@@ -1,0 +1,3 @@
+import type { GrpcWebOptions } from '@protobuf-ts/grpcweb-transport'
+
+export type GrpcWebTransportAdditionalOptions = Omit<GrpcWebOptions, 'baseUrl'>
